Cosmetic: Rename listItems structure.
To reflect the importance of the listItems structure and the fact that it
holds almost all information concerning the GUI, it is renamed 'guiItems' now.

Additionally, the listItems pointer used to collect the items defined in the
skin file is renamed 'skin' now.

That way, the intended purpose should be more obvious and the code in skin.c
is a lot easier to read.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@33024 b3059339-0415-0410-9bf9-f77b7e298cf2