Fix GUI icon bug.
The GUI icon did not display properly but showed up with various distortions.
The reason was the icon mask which hadn't been put to the X server yet when
used.

The icon itself was okay, but is rendered now in a way that doesn't need a
drawable which spares creating a GTK window and destroying it right after.

The locally used GDK variables have been moved inside the function where
they are needed.

Patch with grateful support by Steaphan Greene, sgreene cs.binghamton edu.

This closes Bugzilla #582.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@33614 b3059339-0415-0410-9bf9-f77b7e298cf2