Ensure that filename related config will always be loaded.
Currently, filename related config will only be loaded if MPlayer knows
the filename when it gets called, which isn't the case if either MPlayer
is called in slave mode or the GUI is called without file argument. In
either case, the file to be played is known only later.

If filename related config hasn't yet been read, do it after leaving
the idle mode loop.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@34202 b3059339-0415-0410-9bf9-f77b7e298cf2