libdvdcss: Drop #define of snprintf() to _snprintf() if the former is undefined.
This workaround was added for MinGW originally, but nowadays MinGW supports
snprintf() directly so it has become unnecessary.

Furthermore this can play havoc with Cygwin which does not have a _snprintf()
declaration and gives implicit function declaration warnings. Depending on
compiler flags, this can be a fatal error.

This was committed to the libdvdcss Subversion repository as revision 243.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@34267 b3059339-0415-0410-9bf9-f77b7e298cf2