Workaround a bug in Pulse Audio (http://pulseaudio.org/ticket/866) that causes 2 second delay while trying to drain the current stream.
It seems that if we wait for the data to be played before calling
the pa_stream_drain function, it would not block for the whole
duration of its internal buffer, thus avoid the 2 seconds delay.

The code of this workaround is already used in ao_jack and ao_sdl.

Special thanks to Rune Heggtveit <pblc heggtveit n>
who did all the heavy lifting in finding a working solution.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@34652 b3059339-0415-0410-9bf9-f77b7e298cf2