Disable our included tremor by default, it is too buggy.
As a side effect, the ogg demuxer is only enabled if libogg
is available.
This should not matter usually since we use lavf by default,
but it might make a difference for network streams which
still default to demux_ogg since lavf previously had more
issues both with startup time and chained Ogg files.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@35213 b3059339-0415-0410-9bf9-f77b7e298cf2